Actually round rather than truncate for injury durations

+++ MekHQ/src/mekhq/campaign/mod/am/InjuryUtil.java
@@ -242,7 +242,7 @@ public final class InjuryUtil {
             time += Compute.d6();
         }
 
-        time = Math.round(time * mod * p.getAbilityTimeModifier() / 10000);
+        time = (int)Math.round((time * mod * p.getAbilityTimeModifier()) / 10000.0);
         return time;
     }
     
@@ -269,7 +269,7 @@ public final class InjuryUtil {
                 if (roll < Math.max(1, fumbleLimit / 10)) {
                     mistakeXP += c.getCampaignOptions().getMistakeXP();
                     xpGained += mistakeXP;
-                } else if (roll > Math.min(98, 99 - Math.round(99 - critLimt) / 10)) {
+                } else if (roll > Math.min(98, 99 - (int)Math.round((99 - critLimt) / 10.0))) {
                     successXP += c.getCampaignOptions().getSuccessXP();
                     xpGained += successXP;
                 }
